+++ src/ezdxf/entities/dictionary.py
@@ -1,14 +1,17 @@
 # Copyright (c) 2019-2020, Manfred Moitzi
 # License: MIT-License
 # Created: 2019-02-18
-from typing import TYPE_CHECKING, KeysView, ItemsView, Any, Union, Dict, List
-from ezdxf.lldxf.const import SUBCLASS_MARKER, DXFKeyError
-from ezdxf.lldxf.attributes import DXFAttr, DXFAttributes, DefSubclass
+from typing import TYPE_CHECKING, KeysView, ItemsView, Any, Union, Dict
+import logging
+from ezdxf.lldxf import validator
+from ezdxf.lldxf.const import SUBCLASS_MARKER, DXFKeyError, DXFValueError
+from ezdxf.lldxf.attributes import (
+    DXFAttr, DXFAttributes, DefSubclass, RETURN_DEFAULT,
+)
 from ezdxf.audit import AuditError
 from .dxfentity import base_class, SubclassProcessor, DXFEntity
 from .dxfobj import DXFObject
 from .factory import register_entity
-import logging
 
 logger = logging.getLogger('ezdxf')
 
@@ -18,35 +21,49 @@ if TYPE_CHECKING:
 __all__ = ['Dictionary', 'DictionaryWithDefault', 'DictionaryVar']
 
 acdb_dictionary = DefSubclass('AcDbDictionary', {
-    'hard_owned': DXFAttr(280, default=0, optional=True),  # Hard-owner flag.
-    # If set to 1, indicates that elements of the dictionary are to be treated as hard-owned
-    'cloning': DXFAttr(281, default=1),  # Duplicate record cloning flag (determines how to merge duplicate entries):
+    # If set to 1, indicates that elements of the dictionary are to be treated
+    # as hard-owned:
+    'hard_owned': DXFAttr(
+        280, default=0, optional=True,
+        validator=validator.is_integer_bool,
+        fixer=RETURN_DEFAULT,
+    ),
+
+    # Duplicate record cloning flag (determines how to merge duplicate entries):
     # 0 = not applicable
     # 1 = keep existing
     # 2 = use clone
     # 3 = <xref>$0$<name>
     # 4 = $0$<name>
     # 5 = Unmangle name
-
+    'cloning': DXFAttr(
+        281, default=1,
+        validator=validator.is_in_integer_range(0, 6),
+        fixer=RETURN_DEFAULT,
+    ),
     # 3: entry name
-    # 350: entry handle, some DICTIONARY objects have 360 as handle group code, this is accepted by AutoCAD but not
-    # documented by the DXF reference!!! ezdxf replaces 360 codes by 350.
+    # 350: entry handle, some DICTIONARY objects have 360 as handle group code,
+    # this is accepted by AutoCAD but not documented by the DXF reference!
+    # ezdxf replaces group code 360 by 350.
 })
 
 KEY_CODE = 3
 VALUE_CODE = 350
-SEARCH_CODES = (VALUE_CODE, 360)  # some DICTIONARY have 360 handles
+# Some DICTIONARY use group code 360:
+SEARCH_CODES = (VALUE_CODE, 360)
 
 
 @register_entity
 class Dictionary(DXFObject):
     """
-    AutoCAD maintains items such as mline styles and group definitions as objects in dictionaries.
-    Other applications are free to create and use their own dictionaries as they see fit. The prefix "ACAD_" is reserved
+    AutoCAD maintains items such as mline styles and group definitions as
+    objects in dictionaries. Other applications are free to create and use
+    their own dictionaries as they see fit. The prefix "ACAD_" is reserved
     for use by AutoCAD applications.
 
-    Dictionary entries are (key, DXFEntity) pairs. DXFEntity could be a string, because at loading time not all objects
-    are already stored in the EntityDB, and have to acquired later.
+    Dictionary entries are (key, DXFEntity) pairs. DXFEntity could be a string,
+    because at loading time not all objects are already stored in the EntityDB,
+    and have to acquired later.
 
     """
     DXFTYPE = 'DICTIONARY'
@@ -54,12 +71,13 @@ class Dictionary(DXFObject):
 
     def __init__(self, doc: 'Drawing' = None):
         super().__init__(doc)
-        self._data = dict()  # type: Dict[str, Union[str, DXFEntity]]
-        self._value_code = VALUE_CODE  # some dict have 360 handles for values
+        self._data: Dict[str, Union[str, DXFEntity]] = dict()
+        self._value_code = VALUE_CODE
 
     def _copy_data(self, entity: 'Dictionary') -> None:
-        """ Copy hard owned entities but do not store the copies in the entity database, this is a
-        second step, this is just real copying.
+        """ Copy hard owned entities but do not store the copies in the entity
+        database, this is a second step, this is just real copying.
+
         """
         # todo: what about reactors of cloned DXF objects?
         entity._value_code = self._value_code
@@ -69,7 +87,9 @@ class Dictionary(DXFObject):
             entity._data = {key: entity for key, entity in self.items()}
 
     def _add_data_to_db(self) -> None:
-        """ Add hard owned and therefore copied entities into database and the objects section.  """
+        """ Add hard owned and therefore copied entities into database and the
+        objects section.
+        """
         # todo: don't know how to proceed with reactors of cloned objects?
         if self.dxf.hard_owned:
             my_handle = self.dxf.handle
@@ -77,13 +97,17 @@ class Dictionary(DXFObject):
                 entity.dxf.owner = my_handle
                 entity.dxf.handle = None
                 self.entitydb.add(entity)
-                # add it also to the objects section, else it wouldn't exported to DXF file
+                # Add sub-entities also to the objects section, else they
+                # wouldn't be exported to DXF:
                 self.doc.objects.add_object(entity)
 
-    def load_dxf_attribs(self, processor: SubclassProcessor = None) -> 'DXFNamespace':
+    def load_dxf_attribs(self,
+                         processor: SubclassProcessor = None) -> 'DXFNamespace':
         dxf = super().load_dxf_attribs(processor)
         if processor:
-            tags = processor.load_dxfattribs_into_namespace(dxf, acdb_dictionary)
+            tags = processor.load_dxfattribs_into_namespace(
+                dxf, acdb_dictionary
+            )
             self.load_dict(tags)
         return dxf
 
@@ -93,8 +117,9 @@ class Dictionary(DXFObject):
         value_code = VALUE_CODE
         for code, value in tags:
             if code in SEARCH_CODES:
-                # first store handles, because at this point, not all objects are stored in the EntityDB,
-                # at access convert the handle to DXFEntity
+                # First store handles, because at this point, NOT all objects
+                # are stored in the EntityDB, at first access convert the handle
+                # to a DXFEntity object.
                 value_code = code
                 entry_handle = value
             elif code == KEY_CODE:
@@ -103,16 +128,17 @@ class Dictionary(DXFObject):
                 try:
                     entity = self.entitydb[entry_handle]
                 except KeyError:
-                    entity = entry_handle  # store entity as handle string
+                    # Store entity as handle string:
+                    entity = entry_handle
 
                 self._data[dict_key] = entity
                 entry_handle = None
                 dict_key = None
-        self._value_code = value_code  # use same value code as loaded
+        # Use same value code as loaded:
+        self._value_code = value_code
 
     def export_entity(self, tagwriter: 'TagWriter') -> None:
         """ Export entity specific data as DXF tags. """
-        # base class export is done by parent class
         super().export_entity(tagwriter)
 
         tagwriter.write_tag2(SUBCLASS_MARKER, acdb_dictionary.name)
@@ -122,23 +148,29 @@ class Dictionary(DXFObject):
     def export_dict(self, tagwriter: 'TagWriter'):
         # key: dict key string
         # value: DXFEntity or handle as string
-        # Ignore invalid handles at export, because removing can create an empty dictionary, which is more a problem for
-        # AutoCAD than invalid handles, and removing the whole dictionary is also a problem maybe.
+        # Ignore invalid handles at export, because removing can create an empty
+        # dictionary, which is more a problem for AutoCAD than invalid handles,
+        # and removing the whole dictionary is maybe also a problem.
         for key, value in self._data.items():
             tagwriter.write_tag2(KEY_CODE, key)
-            # value can be a handle string or a DXFEntity
+            # Value can be a handle string or a DXFEntity object:
             if isinstance(value, DXFEntity):
                 if value.is_alive:
                     value = value.dxf.handle
-                else:  # entry was deleted externally
-                    logger.debug(f'Key "{key}" references (external) deleted entry in {str(self)}')
+                else:
+                    logger.debug(
+                        f'Key "{key}" points to a destroyed entity '
+                        f'in {str(self)}, target replaced by "0" handle.'
+                    )
                     value = '0'
-            tagwriter.write_tag2(self._value_code, value)  # use same value code as loaded
+            # Use same value code as loaded:
+            tagwriter.write_tag2(self._value_code, value)
 
     @property
     def is_hard_owner(self) -> bool:
-        """ ``True`` if :class:`Dictionary` is hard owner of entities. Hard owned entities will be deleted by deleting
-        the dictionary.
+        """ ``True`` if :class:`Dictionary` is hard owner of entities. Hard
+        owned entities will be deleted by deleting the dictionary.
+
         """
         return bool(self.dxf.hard_owned)
 
@@ -147,12 +179,18 @@ class Dictionary(DXFObject):
         return self._data.keys()
 
     def items(self) -> ItemsView:
-        """ Returns :class:`ItemsView` for all dictionary entries as (:attr:`key`, :class:`DXFEntity`) pairs. """
+        """ Returns :class:`ItemsView` for all dictionary entries as
+        (:attr:`key`, :class:`DXFEntity`) pairs.
+
+        """
         for key in self.keys():
             yield key, self.get(key)  # maybe handle -> DXFEntity
 
     def __getitem__(self, key: str) -> 'DXFEntity':
-        """ Return the value for `key`, raises a :class:`DXFKeyError` if `key` does not exist. """
+        """ Return the value for `key`, raises a :class:`DXFKeyError` if `key`
+        does not exist.
+
+        """
         return self.get(key)
 
     def __setitem__(self, key: str, value: 'DXFEntity') -> None:
@@ -160,7 +198,10 @@ class Dictionary(DXFObject):
         return self.add(key, value)
 
     def __delitem__(self, key: str) -> None:
-        """ Delete entry `key` from the dictionary, raises :class:`DXFKeyError` if key does not exist. """
+        """ Delete entry `key` from the dictionary, raises :class:`DXFKeyError`
+        if key does not exist.
+
+        """
         return self.remove(key)
 
     def __contains__(self, key: str) -> bool:
@@ -175,22 +216,22 @@ class Dictionary(DXFObject):
 
     def get(self, key: str, default: Any = DXFKeyError) -> 'DXFEntity':
         """
-        Returns :class:`DXFEntity` for `key`, if `key` exist, else `default` or raises a :class:`DXFKeyError`
-        for `default` = :class:`DXFKeyError`.
+        Returns :class:`DXFEntity` for `key`, if `key` exist, else `default` or
+        raises a :class:`DXFKeyError` for `default` = :class:`DXFKeyError`.
 
         """
         try:
             entity = self._data[key]
         except KeyError:
             if default is DXFKeyError:
-                raise DXFKeyError("KeyError: '{}'".format(key))
+                raise DXFKeyError(f"KeyError: '{key}'")
             else:
                 return default
         else:
             if isinstance(entity, str):
-                # entity is still a handle, get DXFEntity from EntityDB
+                # Entity is still a handle:
                 entity = self.entitydb[entity]
-                # and store DXFEntity in the dict
+                # and replace handle by DXFEntity object:
                 self._data[key] = entity
             return entity
 
@@ -200,13 +241,13 @@ class Dictionary(DXFObject):
             try:
                 value = self.entitydb[value]
             except KeyError:
-                raise DXFKeyError('Invalid entity handle #{} for key {}'.format(value, key))
+                raise DXFValueError(
+                    f'Invalid entity handle #{value} for key {key}')
         self._data[key] = value
 
     def remove(self, key: str) -> None:
-        """
-        Delete entry `key`. Raises :class:`DXFKeyError`, if `key` does not exist. Deletes also hard owned DXF
-        objects from OBJECTS section.
+        """ Delete entry `key`. Raises :class:`DXFKeyError`, if `key` does not
+        exist. Deletes also hard owned DXF objects from OBJECTS section.
 
         """
         data = self._data
@@ -215,14 +256,14 @@ class Dictionary(DXFObject):
 
         if self.is_hard_owner:
             entity = self.get(key)
-            # Presumption: hard owned DXF objects always reside in the OBJECTS section
+            # Presumption: hard owned DXF objects always reside in the OBJECTS
+            # section.
             self.doc.objects.delete_entity(entity)
         del data[key]
 
     def discard(self, key: str) -> None:
-        """
-        Delete entry `key` if exists. Does NOT raise an exception if `key` not exist and does not delete hard
-        owned DXF objects.
+        """ Delete entry `key` if exists. Does NOT raise an exception if `key`
+        not exist and does not delete hard owned DXF objects.
 
         """
         try:
@@ -231,7 +272,10 @@ class Dictionary(DXFObject):
             pass
 
     def clear(self) -> None:
-        """  Delete all entries from DXFDictionary, deletes hard owned DXF objects from OBJECTS section. """
+        """  Delete all entries from :class:`Dictionary`, deletes hard owned
+        DXF objects from OBJECTS section.
+
+        """
         if self.is_hard_owner:
             self._delete_hard_owned_entries()
         self._data.clear()
@@ -251,7 +295,8 @@ class Dictionary(DXFObject):
             hard_owned: entries of the new dictionary are hard owned
 
         """
-        dxf_dict = self.doc.objects.add_dictionary(owner=self.dxf.handle, hard_owned=hard_owned)
+        dxf_dict = self.doc.objects.add_dictionary(owner=self.dxf.handle,
+                                                   hard_owned=hard_owned)
         self.add(key, dxf_dict)
         return dxf_dict
 
@@ -263,12 +308,17 @@ class Dictionary(DXFObject):
              value: entry value as string
 
         """
-        new_var = self.doc.objects.add_dictionary_var(owner=self.dxf.handle, value=value)
+        new_var = self.doc.objects.add_dictionary_var(
+            owner=self.dxf.handle,
+            value=value
+        )
         self.add(key, new_var)
         return new_var
 
     def get_required_dict(self, key: str) -> 'Dictionary':
-        """ Get entry `key` or create a new :class:`Dictionary`, if `Key` not exit. """
+        """ Get entry `key` or create a new :class:`Dictionary`,
+        if `Key` not exit.
+        """
         try:
             dxf_dict = self.get(key)
         except DXFKeyError:
@@ -280,19 +330,19 @@ class Dictionary(DXFObject):
         self._check_invalid_entries(auditor)
 
     def _check_invalid_entries(self, auditor: 'Auditor'):
-        keys_to_discard = []  # do not delete content while iterating
-        append = keys_to_discard.append
+        trash = []  # do not delete content while iterating
+        append = trash.append
         db = self.entitydb
         for key, entry in self._data.items():
-            if isinstance(entry, str):  # entry is a handle as string
+            if isinstance(entry, str):
                 if entry not in db:
                     append(key)
-            elif entry.is_alive:  # entry is a DXFEntity and alive
+            elif entry.is_alive:
                 if entry.dxf.handle not in db:
                     append(key)
-            else:  # entry is a DXFEntity and was deleted externally
+            else:  # entry is destroyed
                 append(key)
-        for key in keys_to_discard:
+        for key in trash:
             del self._data[key]
             auditor.fixed_error(
                 code=AuditError.INVALID_DICTIONARY_ENTRY,
@@ -314,13 +364,15 @@ acdb_dict_with_default = DefSubclass('AcDbDictionaryWithDefault', {
 @register_entity
 class DictionaryWithDefault(Dictionary):
     DXFTYPE = 'ACDBDICTIONARYWDFLT'
-    DXFATTRIBS = DXFAttributes(base_class, acdb_dictionary, acdb_dict_with_default)
+    DXFATTRIBS = DXFAttributes(base_class, acdb_dictionary,
+                               acdb_dict_with_default)
 
     def __init__(self, doc: 'Drawing' = None):
         super().__init__(doc)
         self._default = None  # type: DXFEntity
 
-    def load_dxf_attribs(self, processor: SubclassProcessor = None) -> 'DXFNamespace':
+    def load_dxf_attribs(self,
+                         processor: SubclassProcessor = None) -> 'DXFNamespace':
         dxf = super().load_dxf_attribs(processor)
         if processor is None:
             return dxf
@@ -335,9 +387,8 @@ class DictionaryWithDefault(Dictionary):
 
     def get(self, key: str, default: Any = DXFKeyError) -> DXFEntity:
         # `default` argument is ignored, exist only for API compatibility,
-        """
-        Returns :class:`DXFEntity` for `key` or the predefined dictionary wide :attr:`dxf.default`
-        entity if `key` does not exist.
+        """ Returns :class:`DXFEntity` for `key` or the predefined dictionary
+        wide :attr:`dxf.default` entity if `key` does not exist.
 
         """
         if self._default is None:
@@ -359,7 +410,8 @@ class DictionaryWithDefault(Dictionary):
 
 
 acdb_dict_var = DefSubclass('DictionaryVariables', {
-    'schema': DXFAttr(280, default=0),  # Object schema number (currently set to 0)
+    'schema': DXFAttr(280, default=0),
+    # Object schema number (currently set to 0)
     'value': DXFAttr(1, default=''),
 })
 
@@ -367,8 +419,9 @@ acdb_dict_var = DefSubclass('DictionaryVariables', {
 @register_entity
 class DictionaryVar(DXFObject):
     """
-    DICTIONARYVAR objects are used by AutoCAD as a means to store named values in the database for setvar / getvar
-    purposes without the need to add entries to the DXF HEADER section. System variables that are stored as
+    DICTIONARYVAR objects are used by AutoCAD as a means to store named values
+    in the database for setvar / getvar purposes without the need to add entries
+    to the DXF HEADER section. System variables that are stored as
     DICTIONARYVAR objects are the following:
 
         - DEFAULTVIEWCATEGORY
@@ -397,7 +450,8 @@ class DictionaryVar(DXFObject):
     DXFTYPE = 'DICTIONARYVAR'
     DXFATTRIBS = DXFAttributes(base_class, acdb_dict_var)
 
-    def load_dxf_attribs(self, processor: SubclassProcessor = None) -> 'DXFNamespace':
+    def load_dxf_attribs(self,
+                         processor: SubclassProcessor = None) -> 'DXFNamespace':
         dxf = super().load_dxf_attribs(processor)
         if processor is None:
             return dxf